app/requirements.txt
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
waitress==1.1.0
Werkzeug==0.12.2
flask-cors==3.0.3
gitpython==2.1.7
